fix: TitleButtonGroup as seperate component
